# controller/service.go
package controller

import (
	"github.com/e421083458/go_gateway_demo/dto"
	"github.com/e421083458/go_gateway_demo/middleware"
	"github.com/gin-gonic/gin"
)



type ServiceController struct {}

func ServiceRegister(group *gin.RouterGroup){
	service:=&ServiceController{}
	group.GET("/service_list",service.ServiceList)


}

// AdminLogin godoc
// @Summary 服务列表
// @Description 服务列表
// @Tags 服务管理
// @ID /service/service_list
// @Accept  json
// @Produce  json
// @Param info query string false "关键词"
// @Param page_size query int true "每页个数"
// @Param page_no query int true "关键词"
// @Success 200 {object} middleware.Response{data=dto.ServiceListOutput} "success"
// @Router /service/service_list [get]
func (service *ServiceController)ServiceList(c *gin.Context){
	params:=&dto.ServiceListInput{}
	if err:=params.BindValidParam(c);err!=nil{
		middleware.ResponseError(c,2000,err)
		return
	}

	out := &dto.AdminInfoOutput{}
	middleware.ResponseSuccess(c, out)
}